[TASK] Change SN async task context
Move functions which are using the consecutive time into
the syncronous interrupt. This is needed to enable
a common consecutive timebase on both redundant
processors.

Rework consecutive timebase generation.
Use a 16bit hardware timer to generate a timebase and generate
a UINT64 microsecond counter out of it.

Disable all debug messages which are printed in
IR context now. Rework debug messages to work on
all platforms.

Change-Id: I721189f5e433946a841212610fc707f1f8f267ee
Signed-off-by: Thomas Mair <thomas.mair@br-automation.com>