[FIX] Introduce NVM state field
Add a state field to the SOD storage. This field
enables to invalidate an existing SOD image when
a newer image was downloaded.

This field is necessary as it is not possible to
erase the flash during runtime as this can take
very long. (Overwrite on a flash memory is also
not possible!)

Change-Id: Iaf8adcdd852e9c13ba03b40b66f7613b4a298304
Signed-off-by: Thomas Mair <thomas.mair@br-automation.com>