[FIX] Resolve async. channel locking problem
When disconnecting the network cable for a short time,
the asynchronous channel (CN -> MN) could block.
This prevents asynchronous communication related
to openSAFETY. As a result, the openSAFETY SN is
not able to operate correctly / the openSAFETY
SCM is not able to reintegrate the SN for a period
of time. To prevent this behaviour, the SDO channel
is freed, when the OPLK status changes to the state
NMT_CS_PRE_OPERATIONAL_1.

Resolves: #923

Change-Id: I5533ec72ce614baa4c170c6a451cd57b41d17759