[TASK] Process only valid openSAFETY SPDO frames
Ensure that only valid openSAFETY SPDO frames get
processed by checking the first two bytes of the frame
to reduce stack errors. This is useful in situations
having a running openSAFETY network and doing
a power-cycle of the SCM. Because of the POWERLINK
mapping, the SN would get zeroed frames until the
SCM is up and running again.

Change-Id: Ie448d4a3d676dec1056c260b271188dc0555d86f