[FIX] Fix openSAFETY boot issue due to IP buffer allocation failure
There was not sufficient memory for the ARP frame to be stored,
hence the ARP query packets were not sent to the network which
prevented the boot of the openSAFETY demo. To resolve this issue
the vEth receive descriptor was reverted to use individual length
and data fields instead of the ip_packet_type structure. In this
case, the data field size is 1500 bytes which is sufficient for
the receive buffer.

Change-Id: I13922c2f16573b5673523b11f4a59e0125d34a26